move health2 back to health
